chore: some code refactor
