+++ src/api/factory.ts
@@ -1,5 +1,5 @@
 import type {ApolloClientWrapper} from './apollo'
-import {ApiChatFolders, ApiMedia, ApiMessages} from './methods'
+import {ApiMedia, ApiMessages} from './methods'
 import {ApiAccount} from './methods/account'
 import {ApiAuth} from './methods/auth'
 import {ApiChats} from './methods/chats'
@@ -42,10 +42,6 @@ export class ServiceFactory {
 
   public createMessages() {
     return new ApiMessages(this.apolloClient.getClient())
-  }
-
-  public createChatFolders() {
-    return new ApiChatFolders(this.apolloClient.getClient())
   }
 
   public createMedia() {
