+++ src/api/methods/index.ts
@@ -6,5 +6,4 @@ export * from './langPack'
 export * from './settings'
 export * from './users'
 export * from './messages'
-export * from './folders'
 export * from './media'
