--- src/common/errors.ts
@@ -1,3 +1,0 @@
-/**
- * @todo auth and debug error here
- */
