+++ src/components/popups/NewContactModal.scss
@@ -63,6 +63,14 @@
     display: flex;
     gap: 10px;
     margin-bottom: 15px;
+
+    // @include responsive(mobile) {
+    //   align-items: center;
+    //   > * {
+    //     align-items: stretch;
+    //   }
+    //   // width: 100%;
+    // }
   }
 
   &-footer {
