+++ src/containers/left/LeftHeader.tsx
@@ -1,6 +1,6 @@
 import {type FC, type PropsWithChildren, useCallback} from 'preact/compat'
 
-import {LeftColumnScreen} from 'types/ui'
+import {LeftColumnScreen} from 'types/screens'
 
 import {ColumnHeader} from 'components/ColumnHeader'
