+++ src/css/_partials.scss
@@ -32,6 +32,9 @@
 .text-center {
   text-align: center;
 }
+.text-end {
+  text-align: end;
+}
 
 .primary-color {
   color: var(--c-primary);
