+++ src/lib/lottie/LottiePlayer.scss
@@ -47,7 +47,7 @@
   }
   &-fallback {
     // width: max-content;
-    background-color: red;
+    // background-color: red;
     display: flex;
     justify-content: center;
     @include lottie-sizes();
