+++ src/modules/App.scss
@@ -1,7 +1,5 @@
 .auth-page {
   height: calc(100% + 100px);
-  // max-width: 400px;
-  // width: 400px;
   width: 100%;
   position: relative;
   padding: 50px 15px;
