+++ src/modules/auth/Auth.tsx
@@ -7,19 +7,11 @@ import {cleanupUnusedAuthState} from 'state/updates'
 import {AuthScreens} from 'types/screens'
 
 import {Transition} from 'components/transitions'
-import {Button} from 'components/ui'
 
 import AuthCode from './AuthCode.async'
 import AuthPassword from './AuthPassword.async'
 import AuthPhoneNumber from './AuthPhoneNumber'
 import SignUp from './SignUp.async'
-
-const classNames = {
-  // [AuthScreens.Code]: 'Auth_code',
-  [AuthScreens.Password]: 'Auth_password',
-  // [AuthScreens.PhoneNumber]: 'Auth_phone',
-  [AuthScreens.SignUp]: 'Auth_signup',
-}
 
 const Auth: FC = () => {
   const global = getGlobalState()
@@ -45,13 +37,6 @@ const Auth: FC = () => {
   return (
     <div class="scrollable scrollable-y scrollable-hidden" id="auth-scroll">
       <div class="auth-page">
-        <Button
-          onClick={() => {
-            global.auth.screen = AuthScreens.SignUp
-          }}
-        >
-          Sign up
-        </Button>
         <Transition
           // innerClassnames={classNames}
           name="zoomFade"
