+++ src/modules/main/Main.tsx
@@ -1,7 +1,5 @@
 import {type FC, memo, useEffect, useRef} from 'preact/compat'
 
-import {MOCK_STORIES} from 'api/types/stories'
-
 import {getActions} from 'state/action'
 import 'state/actions/all'
 import {type MapState, connect} from 'state/connect'
@@ -10,13 +8,10 @@ import {destroySubscribeAll, subscribeToAll} from 'state/subscribe'
 
 import {useEventListener} from 'hooks/useEventListener'
 
-// import {useBoolean} from 'hooks/useFlag'
 import LeftColumn from 'containers/left/LeftColumn'
 import MiddleColumn from 'containers/middle/MiddleColumn'
 import {RightColumn} from 'containers/right/RightColumn'
-import {StoryViewer} from 'containers/stories/viewer'
 
-// import {StoryViewer} from 'containers/stories/viewer'
 import CommonModal from 'components/popups/CommonModal.async'
 import NewContactModal from 'components/popups/NewContactModal.async'
 import Notification from 'components/popups/Notification.async'
@@ -76,16 +71,6 @@ const Main: FC<StateProps> = ({isNewContactModalOpen, newContactUserId}) => {
       <NewContactModal isOpen={isNewContactModalOpen} userId={newContactUserId} />
       <Notification isOpen={global.notification.isOpen} />
       <CommonModal isOpen={global.commonModal.isOpen} />
-      <StoryViewer
-        isOpen={global.stories.isOpen}
-        onClose={() => {
-          global.stories.isOpen = false
-        }}
-        stories={MOCK_STORIES}
-        onAllStoriesEnd={() => {
-          console.log('STORIES ALL END')
-        }}
-      />
     </div>
   )
 }
