+++ src/state/actions/all.ts
@@ -5,7 +5,6 @@ import './api/account'
 import './api/auth'
 import './api/chats'
 import './api/contacts'
-import './api/folders'
 import './api/langPack'
 import './api/media'
 import './api/messages'
