--- src/state/actions/api/folders.ts
@@ -1,9 +1,0 @@
-import {Api} from 'api/manager'
-
-import {createAction} from 'state/action'
-
-createAction('getChatFolders', async (state, actions, payload) => {
-  const result = await Api.folders.getChatFolders()
-
-  console.log({result})
-})
