+++ src/types/lib.ts
@@ -1,5 +1,4 @@
 import type lang from 'lib/i18n/lang'
-import type errors from 'lib/i18n/errors'
 
 export type ApiLangCode = 'en' | 'uk' | 'pl' | 'de'
 export type SupportedLanguagesList = Array<{
@@ -7,7 +6,5 @@ export type SupportedLanguagesList = Array<{
   label: string
 }>
 export type LanguagePack = typeof lang
-export type ErrorPack = typeof errors
 
 export type LanguagePackKeys = keyof LanguagePack
-export type ErrorPackKeys = keyof ErrorPack
